extracting content hover code into separate classes
