+++ src/vs/editor/contrib/hover/browser/contentHoverComputer.ts
@@ -1,0 +1,110 @@
+/*---------------------------------------------------------------------------------------------
+ *  Copyright (c) Microsoft Corporation. All rights reserved.
+ *  Licensed under the MIT License. See License.txt in the project root for license information.
+ *--------------------------------------------------------------------------------------------*/
+
+import { coalesce } from 'vs/base/common/arrays';
+import { CancellationToken } from 'vs/base/common/cancellation';
+import { IActiveCodeEditor, ICodeEditor } from 'vs/editor/browser/editorBrowser';
+import { IModelDecoration } from 'vs/editor/common/model';
+import { HoverStartSource, IHoverComputer } from 'vs/editor/contrib/hover/browser/hoverOperation';
+import { HoverAnchor, HoverAnchorType, IEditorHoverParticipant, IHoverPart } from 'vs/editor/contrib/hover/browser/hoverTypes';
+import { AsyncIterableObject } from 'vs/base/common/async';
+
+export class ContentHoverComputer implements IHoverComputer<IHoverPart> {
+
+	private _anchor: HoverAnchor | null = null;
+	public get anchor(): HoverAnchor | null { return this._anchor; }
+	public set anchor(value: HoverAnchor | null) { this._anchor = value; }
+
+	private _shouldFocus: boolean = false;
+	public get shouldFocus(): boolean { return this._shouldFocus; }
+	public set shouldFocus(value: boolean) { this._shouldFocus = value; }
+
+	private _source: HoverStartSource = HoverStartSource.Mouse;
+	public get source(): HoverStartSource { return this._source; }
+	public set source(value: HoverStartSource) { this._source = value; }
+
+	private _insistOnKeepingHoverVisible: boolean = false;
+	public get insistOnKeepingHoverVisible(): boolean { return this._insistOnKeepingHoverVisible; }
+	public set insistOnKeepingHoverVisible(value: boolean) { this._insistOnKeepingHoverVisible = value; }
+
+	constructor(
+		private readonly _editor: ICodeEditor,
+		private readonly _participants: readonly IEditorHoverParticipant[]
+	) {
+	}
+
+	private static _getLineDecorations(editor: IActiveCodeEditor, anchor: HoverAnchor): IModelDecoration[] {
+		if (anchor.type !== HoverAnchorType.Range && !anchor.supportsMarkerHover) {
+			return [];
+		}
+
+		const model = editor.getModel();
+		const lineNumber = anchor.range.startLineNumber;
+
+		if (lineNumber > model.getLineCount()) {
+			// invalid line
+			return [];
+		}
+
+		const maxColumn = model.getLineMaxColumn(lineNumber);
+
+		return editor.getLineDecorations(lineNumber).filter((d) => {
+			if (d.options.isWholeLine) {
+				return true;
+			}
+
+			const startColumn = (d.range.startLineNumber === lineNumber) ? d.range.startColumn : 1;
+			const endColumn = (d.range.endLineNumber === lineNumber) ? d.range.endColumn : maxColumn;
+
+			if (d.options.showIfCollapsed) {
+				// Relax check around `showIfCollapsed` decorations to also include +/- 1 character
+				if (startColumn > anchor.range.startColumn + 1 || anchor.range.endColumn - 1 > endColumn) {
+					return false;
+				}
+			} else {
+				if (startColumn > anchor.range.startColumn || anchor.range.endColumn > endColumn) {
+					return false;
+				}
+			}
+
+			return true;
+		});
+	}
+
+	public computeAsync(token: CancellationToken): AsyncIterableObject<IHoverPart> {
+		const anchor = this._anchor;
+
+		if (!this._editor.hasModel() || !anchor) {
+			return AsyncIterableObject.EMPTY;
+		}
+
+		const lineDecorations = ContentHoverComputer._getLineDecorations(this._editor, anchor);
+
+		return AsyncIterableObject.merge(
+			this._participants.map((participant) => {
+				if (!participant.computeAsync) {
+					return AsyncIterableObject.EMPTY;
+				}
+				return participant.computeAsync(anchor, lineDecorations, token);
+			})
+		);
+	}
+
+	public computeSync(): IHoverPart[] {
+		if (!this._editor.hasModel() || !this._anchor) {
+			return [];
+		}
+
+		const lineDecorations = ContentHoverComputer._getLineDecorations(this._editor, this._anchor);
+
+		let result: IHoverPart[] = [];
+		for (const participant of this._participants) {
+			result = result.concat(participant.computeSync(this._anchor, lineDecorations));
+		}
+
+		return coalesce(result);
+	}
+}
+
